refactor: modulize enclave agent
Separate services from the program entry.
