+++ .gitignore
@@ -1,0 +1,5 @@
+*/target
+src/*/target
+src/*/Cargo.lock
+src/libs/protocols/src/image.rs
+src/libs/protocols/src/image_ttrpc.rs
